chore: BE 폴더 복수명 통일
