+++ backend/src/apis/books/books.controller.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from 'express';
 
-import bookService from '@apis/books/books.service';
+import booksService from '@apis/books/books.service';
 
 const getBook = async (req: Request, res: Response) => {
   const bookId = Number(req.params.bookId);
-  const bookData = await bookService.getBookData(bookId);
+  const bookData = await booksService.getBookData(bookId);
 
   // 쿠키에서 유저아이디 검출 미들웨어 추가 -> req.userId -> 북마크 확인
