+++ backend/src/apis/images/images.interface.ts
