+++ backend/src/apis/index.ts
@@ -2,13 +2,11 @@ import { Router } from 'express';
 
 import multer from 'multer';
 
-import articleController from '@apis/article/article.controller';
+import articlesController from '@apis/articles/articles.controller';
 import authController from '@apis/auth/auth.controller';
-import bookController from '@apis/books/books.controller';
-import imageController from '@apis/image/image.controller';
+import booksController from '@apis/books/books.controller';
+import imagesController from '@apis/images/images.controller';
 import { catchAsync } from '@utils/catch-async';
-
-import articlesController from './articles/articles.controller';
 
 const router = Router();
 
@@ -16,14 +14,14 @@ router.post('/auth/signin/local', catchAsync(authController.signIn));
 router.post('/auth/signin/github', catchAsync(authController.signInGithub));
 router.post('/auth/signup', catchAsync(authController.signUp));
 
-router.get('/articles/temporary/:userId', catchAsync(articleController.getTemporaryArticle));
-router.post('/articles/temporary', catchAsync(articleController.saveTemporaryArticle));
-router.post('/image', multer().single('image'), catchAsync(imageController.createImage));
+router.get('/articles/temporary/:userId', catchAsync(articlesController.getTemporaryArticle));
+router.post('/articles/temporary', catchAsync(articlesController.saveTemporaryArticle));
+router.get('/articles/:articleId', catchAsync(articlesController.getArticle));
+router.post('/articles', catchAsync(articlesController.publish));
 
-router.post('/articles', catchAsync(articleController.publish));
+router.post('/image', multer().single('image'), catchAsync(imagesController.createImage));
 
 router.get('/books/search');
-router.get('/books/:bookId', catchAsync(bookController.getBook));
+router.get('/books/:bookId', catchAsync(booksController.getBook));
 
-router.get('/articles/:articleId', catchAsync(articlesController.getArticle));
 export default router;
