+++ backend/src/apis/scraps/scraps.interface.ts
